+++ lib/constants.ts
@@ -1,1 +1,2 @@
 export const KONTIST_API_BASE_URL = "https://api.kontist.com";
+export const KONTIST_SUBSCRIPTION_API_BASE_URL = "wss://api.kontist.com";