+++ lib/graphql/client.ts
@@ -1,19 +1,39 @@
 import { GraphQLClient as GQLClient } from "graphql-request";
+import { SubscriptionClient } from "subscriptions-transport-ws";
 import { Variables } from "graphql-request/dist/src/types";
+import * as ws from "ws";
 
 import { Auth } from "../auth";
-import { RawQueryResponse } from "./types";
+import { RawQueryResponse, SubscriptionType, Subscription } from "./types";
 import { serializeGraphQLError } from "../utils";
 import { GraphQLError, UserUnauthorizedError } from "../errors";
+import { GraphQLClientOpts } from "../types";
+
+type Subscriptions = {
+  [key: number]: {
+    id: number;
+    query: string;
+    type: SubscriptionType;
+    onNext: Function;
+    onError?: Function;
+    unsubscribe: () => void;
+  };
+};
 
 export class GraphQLClient {
-  private client: GQLClient;
   private auth: Auth;
+  private client: GQLClient;
+  private subscriptionEndpoint: string;
+  private subscriptionClient?: SubscriptionClient | null;
+  private subscriptions: Subscriptions = {};
+  private subscriptionId: number = 0;
 
-  constructor(endpoint: string, auth: Auth) {
-    this.client = new GQLClient(endpoint);
+  constructor({ auth, endpoint, subscriptionEndpoint }: GraphQLClientOpts) {
     this.auth = auth;
+    this.client = new GQLClient(endpoint);
+    this.subscriptionEndpoint = subscriptionEndpoint;
   }
+
   /**
    * Send a raw GraphQL request and return its response.
    */
@@ -35,6 +55,124 @@ export class GraphQLClient {
       return data;
     } catch (error) {
       throw new GraphQLError(serializeGraphQLError(error));
+    }
+  };
+
+  /**
+   * Create a subscription client
+   */
+  private createSubscriptionClient = (): SubscriptionClient => {
+    if (!this.auth.tokenManager.token) {
+      throw new UserUnauthorizedError();
+    }
+
+    const WebSocket = typeof window === "undefined" ? ws : window.WebSocket;
+
+    return new SubscriptionClient(
+      this.subscriptionEndpoint,
+      {
+        lazy: true,
+        connectionParams: {
+          Authorization: `Bearer ${this.auth.tokenManager.token.accessToken}`
+        }
+      },
+      WebSocket
+    );
+  };
+
+  /**
+   * Handle disconnection:
+   *
+   * 1. Refresh access token
+   * 2. Destroy existing subscription client
+   * 3. Resubscribe all previously active subscriptions
+   */
+  private handleDisconnection = async (): Promise<void> => {
+    try {
+      await this.auth.tokenManager.refresh();
+    } catch (error) {
+      Object.values(this.subscriptions).forEach(({ onError }) => {
+        if (typeof onError === "function") {
+          onError(error);
+        }
+      });
+    }
+
+    this.subscriptionClient = null;
+    Object.values(this.subscriptions).forEach(subscription => {
+      const { query, type, onNext, onError, id } = subscription;
+      this.subscribe({
+        query,
+        type,
+        onNext,
+        onError,
+        subscriptionId: id
+      });
+    });
+  };
+
+  /**
+   * Subscribe to a topic and call the respective handler when new data or an error is received
+   */
+  public subscribe = ({
+    query,
+    type,
+    onError,
+    onNext,
+    subscriptionId
+  }: {
+    query: string;
+    type: SubscriptionType;
+    onNext: Function;
+    onError?: Function;
+    subscriptionId?: number;
+  }): Subscription => {
+    if (!this.subscriptionClient) {
+      this.subscriptionClient = this.createSubscriptionClient();
+
+      this.subscriptionClient.onDisconnected(this.handleDisconnection);
+    }
+
+    const subscription = this.subscriptionClient.request({
+      query
+    });
+
+    const { unsubscribe } = subscription.subscribe({
+      next(response) {
+        onNext(response.data?.[type]);
+      },
+      error(error) {
+        if (typeof onError === "function") {
+          onError(error);
+        }
+      }
+    });
+
+    const id = subscriptionId || (this.subscriptionId += 1);
+    this.subscriptions[id] = {
+      id,
+      query,
+      type,
+      onNext,
+      onError,
+      unsubscribe
+    };
+
+    return {
+      unsubscribe: this.createUnsubscriber(id)
+    };
+  };
+
+  /**
+   * Create an unsubscribe function to be called by the subscriber
+   */
+  private createUnsubscriber = (subscriptionId: number): (() => void) => () => {
+    this.subscriptions[subscriptionId]?.unsubscribe();
+    delete this.subscriptions[subscriptionId];
+
+    if (Object.keys(this.subscriptions).length === 0) {
+      this.subscriptionClient?.close();
+      this.subscriptionClient = null;
     }
   };
 }