+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Introduce a method to subscribe to new transactions (`models.transaction.subscribe`)
 - Introduce methods for fetching transfers (`models.transfer.fetch`, `models.transfer.fetchAll`)
 
 ### Changed
 - *Breaking*: changed transaction async iterator from `models.transaction` to `models.transaction.fetchAll`.
+
+## [0.21.0] - 2019-11-22
+### Changed
+- Type for `Transaction.category` changed from `string` to `TransactionCategory` enum. This is a *breaking* change.
 
 ## [0.20.0] - 2019-11-18
 ### Added